#ENVELOPEDEV44 создание общей библиотеки
- создал общую библиотеку для всех микросервисов (сообщения, ошибки, Result)
- обновил ReadME
- обновил Dockerfile (чтобы ссылалось на инфраструктурные приколы)
