+++ envelope-backend/docker-compose.yml
@@ -22,8 +22,8 @@ services:
     hostname: auth_service
     
     build: 
-      context: src/services/AuthService
-      dockerfile: Dockerfile
+      context: src/
+      dockerfile: services/AuthService/Dockerfile
     
     ports: 
       - '8080:8080'
@@ -59,8 +59,8 @@ services:
     hostname: task_service
 
     build:
-      context: src/services/TaskService
-      dockerfile: Dockerfile
+      context: src/
+      dockerfile: services/TaskService/Dockerfile
     
     environment:
       - ConnectionStrings__TaskEventStore=Host=task_service_db;Port=5432;Database=taskEventStore;Username=postgres;Password=postgres;
@@ -97,8 +97,8 @@ services:
     hostname: courses_service
 
     build:
-      context: src/services/CoursesService
-      dockerfile: Dockerfile
+      context: src/
+      dockerfile: services/CoursesService/Dockerfile
 
     ports:
       - '8082:8082'
