+++ envelope-backend/envelope-backend.sln
@@ -55,6 +55,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "infrastructure", "infrastru
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Envelope.Integration.Tests", "tests\infastructure\Envelope.Integration.Tests\Envelope.Integration.Tests.csproj", "{DCB11D95-887C-429E-BE3E-A0176BCFE560}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Envelope.Common", "src\infrastructure\Envelope.Common\Envelope.Common.csproj", "{14A8E3FB-FFCC-47D8-B147-4D9113DF081E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -129,6 +131,10 @@ Global
 		{DCB11D95-887C-429E-BE3E-A0176BCFE560}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DCB11D95-887C-429E-BE3E-A0176BCFE560}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DCB11D95-887C-429E-BE3E-A0176BCFE560}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14A8E3FB-FFCC-47D8-B147-4D9113DF081E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14A8E3FB-FFCC-47D8-B147-4D9113DF081E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14A8E3FB-FFCC-47D8-B147-4D9113DF081E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14A8E3FB-FFCC-47D8-B147-4D9113DF081E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -158,6 +164,7 @@ Global
 		{14EF9DD7-FC5E-46E2-B34E-FE96509E0556} = {88A686D7-334B-4BBE-89F7-D5013150A892}
 		{6DF1A153-77D6-4D4C-A2C3-5FBC9CADA4E2} = {5C852739-76E2-431E-87CA-40BA08909EB7}
 		{DCB11D95-887C-429E-BE3E-A0176BCFE560} = {6DF1A153-77D6-4D4C-A2C3-5FBC9CADA4E2}
+		{14A8E3FB-FFCC-47D8-B147-4D9113DF081E} = {88A686D7-334B-4BBE-89F7-D5013150A892}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6DF78641-8E68-4B66-AD8A-26DD6D99A988}
