+++ envelope-backend/src/infrastructure/README.md
@@ -1,3 +1,18 @@
+### 1. Envelope.Common
+
+## ENVELOPE.COMMON  - библиотека для общих структур/классов и тд (использующихся во всех сервисах)
+
+Структура
+
+- /Exceptions - общие исключения, которые могут быть встречены во множестве сервисов
+- /Messages - сообщения для обмена через RabbitMq
+- /ResultPatter - паттерн результата (OnSuccess/OnFailure)
+
+
+
+
+### 2. Envelope.Integration
+
 ## ENVELOPE.INTEGRATION - ШИНА СООБЩЕНИЙ ДЛЯ ОБЩЕНИЯ МЕЖДУ МИКРОСЕРВИСАМИ
 
 Библиотека представляет собой реализацию шины сообщений с использованием RabbitMq и библиотеки RabbitMq.Client.
