+++ envelope-backend/src/services/AuthService/AuthService.Application/AuthService.Application.csproj
@@ -14,6 +14,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\..\infrastructure\Envelope.Common\Envelope.Common.csproj" />
       <ProjectReference Include="..\AuthService.Domain\AuthService.Domain.csproj" />
     </ItemGroup>
