+++ envelope-backend/src/services/CoursesService/CoursesService.Application/CoursesService.Application.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\..\infrastructure\Envelope.Common\Envelope.Common.csproj" />
       <ProjectReference Include="..\CoursesService.Domain\CoursesService.Domain.csproj" />
     </ItemGroup>
