--- envelope-backend/src/services/CoursesService/CoursesService.Application/Exceptions/NotFoundException.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-
-namespace CoursesService.Application.Exceptions;
-
-public class NotFoundException : Exception
-{
-    public NotFoundException(MemberInfo type, object key) : base($"{type.Name} with key {key} not found!") { }
-}
