+++ envelope-backend/src/services/CoursesService/CoursesService.Application/Services/CourseBlockService.cs
@@ -1,10 +1,10 @@
-using CoursesService.Application.Common;
-using CoursesService.Application.Exceptions;
+using Envelope.Common.Exceptions;
 using CoursesService.Application.Mapping;
 using CoursesService.Application.Repositories;
 using CoursesService.Application.Requests.CoursBlock;
 using CoursesService.Application.Services.Interfaces;
 using CoursesService.Domain.Entities;
+using Envelope.Common.ResultPattern;
 
 namespace CoursesService.Application.Services;
 
@@ -21,7 +21,7 @@ public class CourseBlockService: ICourseBlockService
         var id = await _repository.AddAsync(block, cancellationToken);
         
         return id == default ?
-            Result<Guid>.OnError(new NotFoundException(typeof(Course), request.CourseId)) :
+            Result<Guid>.OnFailure(new NotFoundException(typeof(Course), request.CourseId)) :
             Result<Guid>.OnSuccess(id);
     }
 
@@ -30,7 +30,7 @@ public class CourseBlockService: ICourseBlockService
         var isDeleted = await _repository.RemoveAsync(id, cancellationToken);
 
         return !isDeleted ?
-            Result<bool>.OnError(new NotFoundException(typeof(Course), id)) :
+            Result<bool>.OnFailure(new NotFoundException(typeof(Course), id)) :
             Result<bool>.OnSuccess(true);
     }
 
@@ -40,7 +40,7 @@ public class CourseBlockService: ICourseBlockService
         var isUpdated = await _repository.UpdateAsync(updatedCourse, cancellationToken);
 
         return !isUpdated ?
-            Result<bool>.OnError(new NotFoundException(typeof(CourseBlock), request.Id)) :
+            Result<bool>.OnFailure(new NotFoundException(typeof(CourseBlock), request.Id)) :
             Result<bool>.OnSuccess(true);
     }
 }
