+++ envelope-backend/src/services/CoursesService/CoursesService.Application/Services/Interfaces/ICourseBlockService.cs
@@ -1,5 +1,5 @@
-using CoursesService.Application.Common;
 using CoursesService.Application.Requests.CoursBlock;
+using Envelope.Common.ResultPattern;
 
 namespace CoursesService.Application.Services.Interfaces;
