--- envelope-backend/src/services/TaskService/TaskService.Application/Exceptions/NotFoundException.cs
@@ -1,6 +1,0 @@
-﻿namespace TaskService.Application.Exceptions;
-
-public class NotFoundException : Exception
-{
-    public NotFoundException(Type type, object key) : base($"{type.Name} with key {key} not found!") { }
-}
