+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Commands/RemoveTask/RemoveTaskCommandHandler.cs
@@ -1,8 +1,8 @@
-﻿using MediatR;
-using TaskService.Application.Common;
+﻿using Envelope.Common.Exceptions;
+using MediatR;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.EventBus.Interfaces;
 using TaskService.Application.EventStore;
-using TaskService.Application.Exceptions;
 using TaskService.Domain.Events;
 
 namespace TaskService.Application.Handlers.Commands.RemoveTask;
