+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Commands/UpdateTask/UpdateTaskCommand.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using TaskService.Application.Common;
+using Envelope.Common.ResultPattern;
 using TaskService.Domain.Enums;
 
 namespace TaskService.Application.Handlers.Commands.UpdateTask;
