+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Queries/GlobalProjectionQueriesHandler/GetAllGlobalProjection/GetAllGlobalProjectionQuery.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using TaskService.Application.Common;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.Responses.GlobalTaskProjections.GetAllGlobalProjections;
 
 namespace TaskService.Application.Handlers.Queries.GlobalProjectionQueriesHandler.GetAllGlobalProjection;
