+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Queries/GlobalProjectionQueriesHandler/GetGlobalProjection/GetGlobalProjectionQueryHandler.cs
@@ -1,5 +1,6 @@
-﻿using MediatR;
-using TaskService.Application.Common;
+﻿using Envelope.Common.Exceptions;
+using MediatR;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.Exceptions;
 using TaskService.Application.Mapping.Responses;
 using TaskService.Application.Repositories.ReadOnlyRepositories;
