+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Queries/SentToCheckProjectionQueryHandler/GetSentToCheckProjection/GetSentToCheckProjectionQuery.cs
@@ -1,6 +1,5 @@
 ﻿using MediatR;
-using TaskService.Application.Common;
-using TaskService.Application.Responses.GlobalTaskProjections.GetGlobalProjection;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.Responses.SentToCheckProjections.GetSentToCheckProjection;
 
 namespace TaskService.Application.Handlers.Queries.SentToCheckProjectionQueryHandler.GetSentToCheckProjection;
