+++ envelope-backend/src/services/TaskService/TaskService.Application/Handlers/Queries/TaskProjectionProjection/GetAllTaskProjection/GetAllTaskProjectionQueryHandler.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using TaskService.Application.Common;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.Mapping.Responses;
 using TaskService.Application.Repositories.ReadOnlyRepositories;
 using TaskService.Application.Responses.TaskProjections.GetAllTaskProjections;
