+++ envelope-backend/src/services/TaskService/TaskService.Application/TaskService.Application.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\infrastructure\Envelope.Common\Envelope.Common.csproj" />
     <ProjectReference Include="..\TaskService.Domain\TaskService.Domain.csproj" />
   </ItemGroup>
