+++ envelope-backend/tests/services/TaskService.Tests/TaskServiceApplication.Tests/CommandHandlerTests/PositiveDefaultCommandHandlersTests.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using TaskService.Application.Common;
+using Envelope.Common.ResultPattern;
 using TaskService.Application.Handlers.Commands.AddTask;
 using TaskService.Application.Handlers.Commands.RefuseTask;
 using TaskService.Application.Handlers.Commands.RemoveTask;
